+++ Changelog.md
@@ -3,6 +3,7 @@
 ## Bug fixes
 
 * avoid posting empty comments. [#3836](https://github.com/diaspora/diaspora/issues/3836)
+* Delegate parent_author to the target of a RelayableRetraction
 
 ## Refactor